Core_and_Images_pipeline/requirements.txt
# Python 3.6.0
beautifulsoup4==4.8.2
bioinfokit==0.8.8
bs4==0.0.1
efficientnet==1.1.0
gpuinfo==1.0.0a6
GPUtil==1.4.0
graphviz==0.13.2
-e git+git@github.com:hyperopt/hyperopt-sklearn.git@310278276008060345d7bfc7cce2e78b608ec549#egg=hpsklearn
hyperopt==0.1.2
imageio==2.5.0
innvestigate==1.0.8
ipdb==0.13.2
keract==4.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-vis==0.4.1
kerasplotlib==0.1.4
lifelines==0.25.6
lightgbm==2.3.1
matplotlib==3.3.3
matplotlib-venn==0.11.5
more-itertools==7.2.0
multiprocess==0.70.11.1
numpy==1.18.5
nvgpu==0.7.0
nvidia-ml-py==375.53.1
opencv-python==4.1.1.26
opt-einsum==3.2.1
pandas==0.25.3
pickleshare==0.7.5
Pillow==8.0.1
pydicom==1.3.0
scikit-image==0.14.2
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.9.0
six==1.12.0
sklearn==0.0
tensorflow-addons==0.10.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
tf-estimator-nightly==1.14.0.dev2019030115
tf-keras-vis==0.3.1
threadpoolctl==2.0.0
utils==1.0.1
virtualenv==15.1.0
xgboost==0.82
